+++ g_utils.c
@@ -373,7 +373,7 @@ char *G_CopyString (char *in)
 {
 	char	*out;
 	
-	out = gi.TagMalloc (strlen(in)+1, TAG_LEVEL);
+	out = gi.TagMalloc ((int)(strlen(in)+1), TAG_LEVEL);
 	strcpy (out, in);
 	return out;
 }
@@ -384,7 +384,7 @@ void G_InitEdict (edict_t *e)
 	e->inuse = true;
 	e->classname = "noclass";
 	e->gravity = 1.0;
-	e->s.number = e - g_edicts;
+	e->s.number = (int)(e - g_edicts);
 }
 
 /*
