+++ p_weapon.c
@@ -181,7 +181,7 @@ void ChangeWeapon (edict_t *ent)
 			i = ((ent->client->pers.weapon->weapmodel & 0xff) << 8);
 		else
 			i = 0;
-		ent->s.skinnum = (ent - g_edicts - 1) | i;
+		ent->s.skinnum = (int)((ent - g_edicts - 1) | i);
 	}
 
 	if (ent->client->pers.weapon && ent->client->pers.weapon->ammo)
@@ -792,7 +792,7 @@ void weapon_grenadelauncher_fire (edict_t *ent)
 	fire_grenade (ent, start, forward, damage, 600, 2.5, radius);
 
 	gi.WriteByte (svc_muzzleflash);
-	gi.WriteShort (ent-g_edicts);
+	gi.WriteShort ((int)(ent-g_edicts));
 	gi.WriteByte (MZ_GRENADE | is_silenced);
 	gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -853,7 +853,7 @@ void Weapon_RocketLauncher_Fire (edict_t *ent)
 
 	// send muzzle flash
 	gi.WriteByte (svc_muzzleflash);
-	gi.WriteShort (ent-g_edicts);
+	gi.WriteShort ((int)(ent-g_edicts));
 	gi.WriteByte (MZ_ROCKET | is_silenced);
 	gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -909,7 +909,7 @@ void Blaster_Fire (edict_t *ent, vec3_t g_offset, int damage, qboolean hyper, in
 
 	// send muzzle flash
 	gi.WriteByte (svc_muzzleflash);
-	gi.WriteShort (ent-g_edicts);
+	gi.WriteShort ((int)(ent-g_edicts));
 	if (hyper)
 		gi.WriteByte (MZ_HYPERBLASTER | is_silenced);
 	else
@@ -1101,7 +1101,7 @@ void Machinegun_Fire (edict_t *ent)
 	fire_bullet (ent, start, forward, damage, kick, DEFAULT_BULLET_HSPREAD, DEFAULT_BULLET_VSPREAD, MOD_MACHINEGUN);
 
 	gi.WriteByte (svc_muzzleflash);
-	gi.WriteShort (ent-g_edicts);
+	gi.WriteShort ((int)(ent-g_edicts));
 	gi.WriteByte (MZ_MACHINEGUN | is_silenced);
 	gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -1245,7 +1245,7 @@ void Chaingun_Fire (edict_t *ent)
 
 	// send muzzle flash
 	gi.WriteByte (svc_muzzleflash);
-	gi.WriteShort (ent-g_edicts);
+	gi.WriteShort ((int)(ent-g_edicts));
 	gi.WriteByte ((MZ_CHAINGUN1 + shots - 1) | is_silenced);
 	gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -1314,7 +1314,7 @@ void weapon_shotgun_fire (edict_t *ent)
 
 	// send muzzle flash
 	gi.WriteByte (svc_muzzleflash);
-	gi.WriteShort (ent-g_edicts);
+	gi.WriteShort ((int)(ent-g_edicts));
 	gi.WriteByte (MZ_SHOTGUN | is_silenced);
 	gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -1373,7 +1373,7 @@ void weapon_supershotgun_fire (edict_t *ent)
 
 	// send muzzle flash
 	gi.WriteByte (svc_muzzleflash);
-	gi.WriteShort (ent-g_edicts);
+	gi.WriteShort ((int)(ent-g_edicts));
 	gi.WriteByte (MZ_SSHOTGUN | is_silenced);
 	gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -1443,7 +1443,7 @@ void weapon_railgun_fire (edict_t *ent)
 
 	// send muzzle flash
 	gi.WriteByte (svc_muzzleflash);
-	gi.WriteShort (ent-g_edicts);
+	gi.WriteShort ((int)(ent-g_edicts));
 	gi.WriteByte (MZ_RAILGUN | is_silenced);
 	gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -1493,7 +1493,7 @@ void weapon_bfg_fire (edict_t *ent)
 	{
 		// send muzzle flash
 		gi.WriteByte (svc_muzzleflash);
-		gi.WriteShort (ent-g_edicts);
+		gi.WriteShort ((int)(ent-g_edicts));
 		gi.WriteByte (MZ_BFG | is_silenced);
 		gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -1600,7 +1600,7 @@ void weapon_ionripper_fire (edict_t *ent)
 
 	// send muzzle flash
 	gi.WriteByte (svc_muzzleflash);
-	gi.WriteShort (ent - g_edicts);
+	gi.WriteShort ((int)(ent - g_edicts));
 	gi.WriteByte (MZ_IONRIPPER | is_silenced);
 	gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -1687,7 +1687,7 @@ void weapon_phalanx_fire (edict_t *ent)
 
 		// send muzzle flash
 		gi.WriteByte (svc_muzzleflash);
-		gi.WriteShort (ent-g_edicts);
+		gi.WriteShort ((int)(ent-g_edicts));
 		gi.WriteByte (MZ_PHALANX | is_silenced);
 		gi.multicast (ent->s.origin, MULTICAST_PVS);
